servidor_html/requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
pandas
pandas-gbq
openpyxl
psutil
playwright
google-cloud-bigquery
jinja2
